fix(agent): share hardened error coercion with TurnEmitter.onError
TurnEmitter.onError evaluated `err instanceof Error` and read
`err.message` unguarded: a revoked Proxy (or any value with a throwing
getPrototypeOf trap / poisoned message getter) threw a TypeError out of
onError, emitted only `start`, and left `stream.result()` unsettled —
despite the class documenting every method as throw-safe.

Extract the adapter's hardened `coerceErrorMessage` into a shared
`provider/error-coercion.ts` and use it from BOTH the stream adapter's
failsafe path and `TurnEmitter.onError` (no duplicated logic). The
adapter's TurnEmitter-independent failsafe remains as defense in depth.

Audit of the remaining public emitter surface found no other unguarded
error-ish reads: onDelta/onFinal already route caller-data throws to the
(now hardened) onError, onAborted takes no value, and finishWithError
only receives plain strings.

Regression test: onError(revokedProxy) on a fresh emitter → no throw,
exactly one error terminal, result() settles. Verified the test fails
against the previous unguarded expression.
